finish partial of the code
Previous version with avatar moving function works but not when I ask Cursor to add new feature to show direction on the map when associated direction is pressed. And I hit the limit of free usage then I just keep the code as it is for now and fix it later by myself.